requirements.txt
numpy
pandas
json
csv
matplotlib.pyplot
flask
datetime